World items only stay for 60 days now, saves some bandwidth.
git-svn-id: http://svn2.xp-dev.com/svn/Fenix1053-MTARP@706 64450a49-1f69-0410-a0a2-f5ebb52c4f5b

+++ src_trunk/resources/item-system/s_item_system.lua
@@ -606,11 +606,11 @@ function loadWorldItems(res)
 			local time = getRealTime()
 			local yearday = time.yearday
 			
-			--[[
-			--if (yearday>(wyearday+30)) then
-				--mysql_query(handler, "DELETE FROM worlditems WHERE id='" .. tonumber(row[1]) .. "' LIMIT 1")
-			--elseif (wyearday>yearday) then -- a new year
-				--mysql_query(handler, "UPDATE worlditems SET yearday='" .. yearday .. "' WHERE id='" .. tonumber(row[1]) .. "' LIMIT 1")
+			
+			if (yearday>(wyearday+60)) then
+				mysql_query(handler, "DELETE FROM worlditems WHERE id='" .. tonumber(row[1]) .. "' LIMIT 1")
+			elseif (wyearday>yearday) then -- a new year
+				mysql_query(handler, "UPDATE worlditems SET yearday='" .. yearday .. "' WHERE id='" .. tonumber(row[1]) .. "' LIMIT 1")
 				
 				local x = tonumber(row[6])
 				local y = tonumber(row[7])
@@ -663,7 +663,6 @@ function loadWorldItems(res)
 					end
 				end
 			else
-			]]--
 				local interior = tonumber(row[10])
 				local dimension = tonumber(row[9])
 				local x = tonumber(row[6])
